pyproject.toml
[tool.poetry]
name = "fsdtorch"
version = "0.1.3"
description = "Face shape detector in pytorch that works with pretrained networks"
authors = ["Alper <itsc0508@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
opencv-contrib-python = "^4.5.3"
onnx = "^1.10.1"
onnxruntime-gpu = "^1.8.1"
gdown = "^3.13.0"
facenet-pytorch = "^2.5.2"

[tool.poetry.dev-dependencies]
poethepoet = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]